design : remove some usages of TextField and LabelledTextField
