+++ features/roomdetails/impl/src/main/kotlin/io/element/android/features/roomdetails/impl/edit/RoomDetailsEditView.kt
@@ -9,7 +9,6 @@
 
 package io.element.android.features.roomdetails.impl.edit
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -21,7 +20,6 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -33,7 +31,6 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import io.element.android.compound.theme.ElementTheme
 import io.element.android.features.roomdetails.impl.R
-import io.element.android.libraries.designsystem.components.LabelledTextField
 import io.element.android.libraries.designsystem.components.async.AsyncActionView
 import io.element.android.libraries.designsystem.components.async.AsyncActionViewDefaults
 import io.element.android.libraries.designsystem.components.avatar.AvatarSize
@@ -45,6 +42,7 @@ import io.element.android.libraries.designsystem.theme.aliasScreenTitle
 import io.element.android.libraries.designsystem.theme.components.Scaffold
 import io.element.android.libraries.designsystem.theme.components.Text
 import io.element.android.libraries.designsystem.theme.components.TextButton
+import io.element.android.libraries.designsystem.theme.components.TextField2
 import io.element.android.libraries.designsystem.theme.components.TopAppBar
 import io.element.android.libraries.matrix.ui.components.AvatarActionBottomSheet
 import io.element.android.libraries.matrix.ui.components.EditableAvatarView
@@ -110,40 +108,28 @@ fun RoomDetailsEditView(
             )
             Spacer(modifier = Modifier.height(60.dp))
 
-            if (state.canChangeName) {
-                LabelledTextField(
-                    label = stringResource(id = R.string.screen_room_details_room_name_label),
-                    value = state.roomRawName,
-                    placeholder = stringResource(CommonStrings.common_room_name_placeholder),
-                    singleLine = true,
-                    onValueChange = { state.eventSink(RoomDetailsEditEvents.UpdateRoomName(it)) },
-                )
-            } else {
-                LabelledReadOnlyField(
-                    title = stringResource(R.string.screen_room_details_room_name_label),
-                    value = state.roomRawName
-                )
-            }
+            TextField2(
+                label = stringResource(id = R.string.screen_room_details_room_name_label),
+                value = state.roomRawName,
+                placeholder = stringResource(CommonStrings.common_room_name_placeholder),
+                singleLine = true,
+                readOnly = !state.canChangeName,
+                onValueChange = { state.eventSink(RoomDetailsEditEvents.UpdateRoomName(it)) },
+            )
 
             Spacer(modifier = Modifier.height(28.dp))
 
-            if (state.canChangeTopic) {
-                LabelledTextField(
-                    label = stringResource(CommonStrings.common_topic),
-                    value = state.roomTopic,
-                    placeholder = stringResource(CommonStrings.common_topic_placeholder),
-                    maxLines = 10,
-                    onValueChange = { state.eventSink(RoomDetailsEditEvents.UpdateRoomTopic(it)) },
-                    keyboardOptions = KeyboardOptions(
-                        capitalization = KeyboardCapitalization.Sentences,
-                    ),
-                )
-            } else {
-                LabelledReadOnlyField(
-                    title = stringResource(R.string.screen_room_details_topic_title),
-                    value = state.roomTopic
-                )
-            }
+            TextField2(
+                label = stringResource(CommonStrings.common_topic),
+                value = state.roomTopic,
+                placeholder = stringResource(CommonStrings.common_topic_placeholder),
+                maxLines = 10,
+                readOnly = !state.canChangeTopic,
+                onValueChange = { state.eventSink(RoomDetailsEditEvents.UpdateRoomTopic(it)) },
+                keyboardOptions = KeyboardOptions(
+                    capitalization = KeyboardCapitalization.Sentences,
+                ),
+            )
         }
     }
 
@@ -169,30 +155,6 @@ fun RoomDetailsEditView(
     PermissionsView(
         state = state.cameraPermissionState,
     )
-}
-
-@Composable
-private fun LabelledReadOnlyField(
-    title: String,
-    value: String,
-) {
-    Column(
-        verticalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        Text(
-            modifier = Modifier.padding(horizontal = 16.dp),
-            style = ElementTheme.typography.fontBodyMdRegular,
-            color = MaterialTheme.colorScheme.primary,
-            text = title,
-        )
-
-        Text(
-            modifier = Modifier.padding(start = 16.dp, end = 16.dp, top = 8.dp, bottom = 12.dp),
-            style = MaterialTheme.typography.bodyLarge,
-            color = MaterialTheme.colorScheme.primary,
-            text = value,
-        )
-    }
 }
 
 @PreviewsDayNight
